[refactor] #22 국회의원 이미지 저장부분 계층 분할
